requirements.txt
discord.py>=2.3.2
docker>=6.1.3
python-dotenv>=1.0.0
aiohttp>=3.8.1
async-timeout>=4.0.2